cut into modules
